Add working example for clinvar scv gk-pilot v2
This is the original working example used to define the v2 pilot clinvar SCV data. It is based on clinvar variation id 208366 and the scv from the expert panel in SCV001245167.1. It has many notes and is likely to be modified significantly before we produce a solid reference example to begin the pilot implementation production.