Added 'qualifier' back to core-im-source.yaml
Larry dropped this in he latest refactor. But IMO defining `qualifier` in the core-im is important for two reasons. It is useful to explain what the notion of a ‘qualifier’ means and what it is used for in the modeling framework. And it is necessary to implement the conceptual specialization I describe above, which  happens when a specific qualifier property is defined in a Statement profile.