Update StudyResult.dataItem property
Here I update StudyResult.dataItem property to align with analogous paradigm for qualifier specialization


In #134 @larrybabb  proposed that profiles can define specific qualifiers as ‘new’ properties as long as they are named as qualifiers.  @mbrush clarified that these are not truly ‘new’ – as conceptually they specialize the core-im `Statement.qualifier` property. 

The use of `StudyResult.dataItem` property in StudyResult profiles is quite analogous, in that profiling often requires defining >1 specialization of this property for different specific data types. 

In this PR, we update the definition of StudyResult.dataItem in the core-im to support this paradigm and align with how `Statement.qualifier` is defined in PR #___

Specifically: 
- removed the existing `dataItems` property
- replaced it with a ‘dataItem’ property (singular, consistent with it not taking an array in the core-im)
- made it take an ‘object’  (which by my understanding covers strings)
- gave it a clear and informative free-text description
- added a ‘comment’ that explains how this property can get profiled into >1 more specific properties. 

In the future we might consider some type of flag/keyword on this core-im property to explicitly mark it as amenable to 1:m specialization in profiles – if this helps with conceptual clarity or computational validation.