Added back InformationEntity, Method, and Document to core-im-source.yaml
Added back `InformationEntity`, `Method`, and `Document` classes (which Larry had moved into gks-commons when we thought we needed them to include RecordMetadata and Contribution classes there.  But updates to these classes no longer requires them to be in gks-commons, so we can move them back here. We can always move them back to gks-commons if we identify broader use cases outside of va-spec for these classes.